chore(e2e): split tasks into dev and prod, put common tasks in e2e
